Backend release

+++ dialob-components/dialob-rule-parser/src/main/java/io/dialob/rule/parser/function/DefaultFunctions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2015 - 2021 ReSys (info@dialob.io)
+ * Copyright © 2015 - 2025 ReSys (info@dialob.io)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -82,13 +82,13 @@ public class DefaultFunctions {
     int[] factors = {7, 9, 10, 5, 8, 4, 2};
     int sum = 0;
     for (int i = 0; i < 7; i++) {
-      sum += factors[i] * (int)(lyt.charAt(i) - '0');
+      sum += factors[i] * (lyt.charAt(i) - '0');
     }
     int reminder = sum % 11;
     if (reminder > 0) {
       reminder = 11 - reminder;
     }
-    return (int)(lyt.charAt(8) - '0') == reminder;
+    return (lyt.charAt(8) - '0') == reminder;
   }
 
   public static boolean isNotHetu(String hetu) {
